+++ experimental/0.1.12/event/dispatcher.js
@@ -1,48 +1,15 @@
 (function (window, undefined) {
     "use strict";
     var ya = window.ya || {},
-        __find = ya.mixins.Array.find,
+        __findAllByFn = ya.mixins.Array.findAllByFn,
         __each = ya.mixins.Array.each,
+        __slice = Array.prototype.slice,
         Dispatcher;
 
     /**
      * @type {Dispatcher}
      */
-    Dispatcher = ya.Core.$extend({
-        defaults: {
-            delegates: null
-        },
-        /**
-         * @param opts
-         * @returns {Dispatcher}
-         */
-        init: function (opts) {
-            // Standard way of initialization.
-            var me = this, config;
-
-            Dispatcher.Parent.init.apply(this, arguments);
-
-            opts = opts || {};
-            config = ya.$merge(me._config, opts.config);
-
-            me.set('initOpts', opts);
-            me.set('config', config);
-
-            me.initConfig();
-
-            return me;
-        },
-        initConfig: function () {
-            var me = this;
-
-            // After calling parent method
-            Dispatcher.Parent.initConfig.apply(this, arguments);
-
-            // set defaults.
-            me.setDelegates([]);
-
-            return me;
-        },
+    Dispatcher = ya.event.Dispatcher.$extend({
         /**
          * @param scope
          * @param e
@@ -50,70 +17,15 @@
          */
         add: function (scope, e) {
             var me = this,
-                query = Object.keys(e).pop(),
-                event = e[query],
-                dGroup = {},
-                pos, view;
+                selector = Object.keys(e).pop();
 
-            query = query.split(" ");
-            view = query.shift();
-            if (view.search(/\$/) === 0) {
-
-                view = view.replace("$", "");
-                query = query.join(" ");
-                pos = me.findGroup(view);
-                if (pos >= 0) {
-
-                    dGroup = me.getGroup(pos);
-                    pos = __find.call(dGroup.selectors, 'selector', query);
-                    if (pos >= 0) {
-                        dGroup.selectors[0].events.push(event);
-                    } else {
-
-                        dGroup.selectors.push({
-                            scope: scope,
-                            selector: query,
-                            events: [
-                                event
-                            ]
-                        });
-
-                    }
-
-                } else {
-
-                    dGroup.viewId = view;
-                    dGroup.selectors = [
-                        {
-                            scope: scope,
-                            selector: query,
-                            events: [
-                                event
-                            ]
-                        }
-                    ];
-
-                    me.getDelegates().push(dGroup);
-
-                }
-
-            }
+            me.getDelegates().push({
+                selector: selector,
+                scope: scope,
+                events: e[selector]
+            });
 
             return me;
-        },
-        /**
-         * @param viewId
-         * @returns {number}
-         */
-        findGroup: function (viewId) {
-            return __find.call(this.getDelegates(), 'viewId', viewId);
-        },
-        /**
-         * @param pos
-         * @returns {*}
-         */
-        getGroup: function (pos) {
-            return this.getDelegates()[pos];
         },
         /**
          * @param view
@@ -121,58 +33,81 @@
         apply: function (view) {
             // Apply delegated events.
             var me = this,
+                delegates = me.getDelegates(),
+                matchPos = [],
+                newView = view,
+                els = [],
                 newScope = function (func, scope, arg) {
                     return func.bind(scope, arg);
                 },
-                els, group;
+                matchIdFn = function (r) {
+
+                    return r.selector.search(view.getId()) >= 0;
+
+                },
+                selector, delegate, e;
 
             while (view) {
 
-                group = me.findGroup(view.getId());
-                if (group >= 0) {
+                matchPos = __findAllByFn.call(delegates, matchIdFn);
+                if (matchPos.length) {
 
-                    group = me.getGroup(group);
-                    __each.call(group.selectors, function (r, i, a) {
+                    __each.call(matchPos, function (r) {
 
-                        if (r.selector)
-                            els = view.queryEls(r.selector);
-                        else
-                            els = [view.get('el')];
+                        els.length = 0;
+                        delegate = delegates[r];
+                        selector = delegate
+                            .selector
+                            .split(" ");
+                        selector.shift();
 
-                        console.log(r.selector, els)
+                        if (selector.length) {
 
-//                        if (els.length) {
-//
-//                            __each.call(els, function (node) {
-//
-//                                __each.call(r.events, function (e) {
-//
-//                                    for (var eType in e) {
-//                                        if (e.hasOwnProperty(eType)) {
-//
-//                                            node.addEventListener(eType, newScope(e[eType], r.scope, view), false);
-//
-//                                        }
-//                                    }
-//
-//                                });
-//
-//                            });
-//
-//                        }
+                            els = __slice.call(newView.queryEls(selector.join(" ")), 0);
+                            if (selector.length === 1 && newView.get('el').nodeName.toLowerCase() === selector[0]) {
+                                els.push(newView.get('el'));
+                            }
+
+                        } else {
+
+                            els.push(view.get('el'));
+
+                        }
+
+                        if (els.length) {
+
+                            __each.call(els, function (node) {
+
+                                e = delegate.events;
+                                for (var eType in e) {
+                                    if (e.hasOwnProperty(eType)) {
+
+                                        node.addEventListener(eType, newScope(e[eType], delegate.scope, view), false);
+
+                                    }
+                                }
+
+                            });
+
+                        }
+
+                        console.log(delegates[r]
+                            .selector, els);
 
                     });
 
                 }
 
+
                 view = view.getParent();
+
             }
 
         }
     });
 
     window.ya = ya;
-    window.ya.event = window.ya.event || {};
-    window.ya.event.Dispatcher = Dispatcher;
-    window.ya.event.dispatcher = Dispatcher.$create();
+    window.ya.experimental = window.ya.experimental || {};
+    window.ya.experimental.event = window.ya.experimental.event || {};
+    window.ya.experimental.event.dispatcher = Dispatcher.$create();
 }(window));